+++ docs/changelog.rst
@@ -57,6 +57,8 @@ Detailed list of changes
 
 - Allow using the cwd of the original process for :option:`launch --cwd` (:iss:`5672`)
 
+- Session files: Expand environment variables (:disc:`5917`)
+
 - Pass key events mapped to scroll actions to the program running in the terminal when the terminal is in alternate screen mode (:iss:`5839`)
 
 - Implement :ref:`edit-in-kitty <edit_file>` using the new ``kitten`` static executable (:iss:`5546`, :iss:`5630`)